# Optimize the consumer assignment of Kafka routine load job