--- a/be/src/common/config.h
+++ b/be/src/common/config.h
@@ -393,6 +393,9 @@ namespace config {
 
     // txn commit rpc timeout
     CONF_Int32(txn_commit_rpc_timeout_ms, "10000");
+
+    // max consumer num in one data consumer group, for routine load
+    CONF_Int32(max_consumer_num_per_group, "1");
 } // namespace config
 
 } // namespace doris