--- a/be/src/runtime/routine_load/data_consumer_group.cpp
+++ b/be/src/runtime/routine_load/data_consumer_group.cpp
@@ -95,7 +95,9 @@ Status KafkaDataConsumerGroup::start_all(StreamLoadContext* ctx) {
                     << ". consume time(ms)=" << ctx->max_interval_s * 1000 - left_time
                     << ", received rows=" << ctx->max_batch_rows - left_rows
                     << ", received bytes=" << ctx->max_batch_size - left_bytes
-                    << ", eos: " << eos;
+                    << ", eos: " << eos
+                    << ", blocking get time(us): " << _queue.total_get_wait_time() / 1000
+                    << ", blocking put time(us): " << _queue.total_put_wait_time() / 1000;
             
             // shutdown queue
             _queue.shutdown();