--- a/be/src/runtime/routine_load/data_consumer_pool.cpp
+++ b/be/src/runtime/routine_load/data_consumer_pool.cpp
@@ -16,6 +16,7 @@
 // under the License.
 
 #include "runtime/routine_load/data_consumer_pool.h"
+#include "runtime/routine_load/data_consumer_group.h"
 
 namespace doris {
 
@@ -60,6 +61,30 @@ Status DataConsumerPool::get_consumer(
     return Status::OK;
 }
 
+Status DataConsumerPool::get_consumer_grp(
+        StreamLoadContext* ctx,
+        std::shared_ptr<DataConsumerGroup>* ret) {
+    if (ctx->load_src_type != TLoadSourceType::KAFKA) {
+        return Status("Currently nly support consumer group for Kafka data source");
+    }
+    DCHECK(ctx->kafka_info);
+
+    std::shared_ptr<KafkaDataConsumerGroup> grp = std::make_shared<KafkaDataConsumerGroup>(ctx);;
+
+    // one data consumer group contains at most 3 data consumers.
+    size_t consumer_num = std::min((size_t) 3, ctx->kafka_info->begin_offset.size());
+    for (int i = 0; i < consumer_num; ++i) {
+        std::shared_ptr<DataConsumer> consumer;
+        RETURN_IF_ERROR(get_consumer(ctx, &consumer));
+        grp->add_consumer(consumer);
+    }
+
+    LOG(INFO) << "get consumer group " << grp->grp_id() << " with "
+            << consumer_num << " consumers";
+    *ret = grp;
+    return Status::OK; 
+}
+
 void DataConsumerPool::return_consumer(std::shared_ptr<DataConsumer> consumer) {
     std::unique_lock<std::mutex> l(_lock);
 
@@ -75,6 +100,12 @@ void DataConsumerPool::return_consumer(std::shared_ptr<DataConsumer> consumer) {
     VLOG(3) << "return the data consumer: " << consumer->id()
             << ", current pool size: " << _pool.size();
     return;
+}
+
+void DataConsumerPool::return_consumers(DataConsumerGroup* grp) {
+    for (std::shared_ptr<DataConsumer> consumer : grp->consumers()) {
+        return_consumer(consumer);
+    }
 }
 
 Status DataConsumerPool::start_bg_worker() {