… be/src/runtime/CMakeLists.txt
… be/src/runtime/routine_load/data_consumer_pool.h
… be/src/runtime/routine_load/routine_load_task_executor.cpp
… be/src/runtime/stream_load/stream_load_context.cpp
… be/src/util/semaphore.hpp
… docs/help/Contents/Data Manipulation/routine_load.md
… fe/src/main/java/org/apache/doris/analysis/CreateRoutineLoadStmt.java
… fe/src/main/java/org/apache/doris/load/LoadChecker.java
… fe/src/main/java/org/apache/doris/load/routineload/KafkaProgress.java
… fe/src/main/java/org/apache/doris/load/routineload/KafkaRoutineLoadJob.java
… fe/src/main/java/org/apache/doris/load/routineload/KafkaTaskInfo.java
… fe/src/main/java/org/apache/doris/load/routineload/RoutineLoadJob.java
… fe/src/main/java/org/apache/doris/load/routineload/RoutineLoadManager.java
… fe/src/main/java/org/apache/doris/load/routineload/RoutineLoadScheduler.java
… fe/src/main/java/org/apache/doris/metric/MetricRepo.java
… fe/src/main/java/org/apache/doris/transaction/GlobalTransactionMgr.java
… fe/src/main/java/org/apache/doris/transaction/PublishVersionDaemon.java
… fe/src/main/java/org/apache/doris/transaction/TransactionState.java
… fe/src/main/java/org/apache/doris/transaction/TxnStateChangeListener.java
… fe/src/test/java/org/apache/doris/load/routineload/KafkaRoutineLoadJobTest.java
… fe/src/test/java/org/apache/doris/load/routineload/RoutineLoadManagerTest.java
… fe/src/test/java/org/apache/doris/transaction/GlobalTransactionMgrTest.java